requirements.txt
requests==2.32.3
celery==5.4.0
PyVimeo==1.1.2
pandas==2.2.2
streamlit==1.38.0
